requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pypi_0
blas=1.0=mkl
blosc=1.21.0=h0e60522_0
brotli=1.0.9=h0e60522_4
bzip2=1.0.8=h8ffe710_4
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py39hcbf5309_1
chardet=4.0.0=pypi_0
charls=2.2.0=h39d44d4_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
configparser=5.0.2=pypi_0
cudatoolkit=10.2.89=hb195166_6
cycler=0.10.0=py_2
cython=0.29.22=pypi_0
cytoolz=0.11.0=py39hb82d6ee_3
dask-core=2021.4.0=pyhd8ed1ab_0
decorator=4.4.2=pypi_0
docker-pycreds=0.4.0=pypi_0
freetype=2.10.4=h546665d_1
fsspec=2021.4.0=pyhd8ed1ab_0
giflib=5.2.1=h8d14728_2
gitdb=4.0.7=pypi_0
gitpython=3.1.14=pypi_0
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.37.0=pypi_0
idna=2.10=pypi_0
imagecodecs=2021.3.31=py39h5da4933_0
imagecorruptions=1.1.2=pypi_0
imageio=2.9.0=py_0
intel-openmp=2020.3=h57928b3_311
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py39h2e07f2f_1
lcms2=2.12=h83e58a3_0
lerc=2.2.1=h0e60522_0
libaec=1.0.4=h39d44d4_1
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libdeflate=1.7=h8ffe710_5
liblapack=3.9.0=8_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.1.0=h56a325e_1
libuv=1.41.0=h8ffe710_0
libzopfli=1.0.3=h0e60522_0
locket=0.2.0=py_2
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pypi_0
matplotlib-base=3.3.4=py39h49ac443_0
mkl=2020.4=hb70f87d_311
mkl-service=2.3.0=py39h4cdbadb_2
mkl_fft=1.3.0=py39h46781fe_0
mkl_random=1.0.2=py39h848d8c7_0
msys2-conda-epoch=20160418=1
networkx=2.5=pypi_0
ninja=1.10.2=h5362a0b_0
numpy=1.20.1=pypi_0
numpy-base=1.19.2=py39hbd0edd7_0
oauthlib=3.1.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.1.48=pypi_0
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1k=h8ffe710_0
partd=1.2.0=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pillow=8.1.2=py39h4fa10fc_0
pip=20.3.3=py39haa95532_0
promise=2.3=pypi_0
protobuf=3.15.8=pypi_0
psutil=5.8.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.9.1=h6244533_2
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pytorch=1.7.1=py3.9_cuda102_cudnn7_0
pywavelets=1.1.1=py39h8512cfa_3
pyyaml=5.4.1=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.1=pypi_0
sentry-sdk=1.0.0=pypi_0
setuptools=49.6.0=py39hcbf5309_3
shortuuid=1.0.1=pypi_0
six=1.15.0=pyh9f0ad1d_0
smmap=4.0.0=pypi_0
snappy=1.1.8=ha925a31_3
sqlite=3.34.0=h8ffe710_0
subprocess32=3.5.4=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.2.1=pypi_0
tk=8.6.10=h8ffe710_1
toolz=0.11.1=py_0
torchaudio=0.7.2=py39
torchvision=0.8.2=py39_cu102
tornado=6.1=py39hb82d6ee_1
typing_extensions=3.7.4.3=py_0
tzdata=2021a=he74cb21_0
urllib3=1.26.4=pypi_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
wandb=0.10.25=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py39hcbf5309_1006
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zfp=0.5.5=h0e60522_5
zlib=1.2.11=h62dcd97_1010
zstd=1.4.5=h04227a9_0
